+++ hutool-core/src/main/java/cn/hutool/core/util/NumberUtil.java
@@ -1889,6 +1889,42 @@ public class NumberUtil {
 	public static boolean isBeside(int number1, int number2) {
 		return Math.abs(number1 - number2) == 1;
 	}
+	
+	/**
+	 * 把给定的总是平均分成N份，返回每份的个数<br>
+	 * 当除以分数有余数时每份+1
+	 * 
+	 * @param total 总数
+	 * @param partCount 份数
+	 * @return 每份的个数
+	 * @since 4.0.7
+	 */
+	public static int partValue(int total, int partCount) {
+		return partValue(total, partCount, true);
+	}
+	
+	/**
+	 * 把给定的总是平均分成N份，返回每份的个数<br>
+	 * 如果isPlusOneWhenHasRem为true，则当除以分数有余数时每份+1，否则丢弃余数部分
+	 * 
+	 * @param total 总数
+	 * @param partCount 份数
+	 * @param isPlusOneWhenHasRem 在有余数时是否每份+1
+	 * @return 每份的个数
+	 * @since 4.0.7
+	 */
+	public static int partValue(int total, int partCount, boolean isPlusOneWhenHasRem) {
+		int partValue = 0;
+		if (total % partCount == 0) {
+			total = total / partCount;
+		} else {
+			partValue = (int) Math.floor(total / partCount);
+			if(isPlusOneWhenHasRem) {
+				partValue += 1;
+			}
+		}
+		return partValue;
+	}
 
 	// ------------------------------------------------------------------------------------------- Private method start
 	private static int mathSubnode(int selectNum, int minNum) {